Move psql_array_to_vector function to apidb
