+++ include/cgimap/backend/apidb/common_pgsql_selection.hpp
@@ -39,4 +39,8 @@ void extract_changesets(
   const std::chrono::system_clock::time_point &now,
   bool include_changeset_discussions);
 
+// parses psql array based on specs given
+// https://www.postgresql.org/docs/current/static/arrays.html#ARRAYS-IO
+std::vector<std::string> psql_array_to_vector(std::string str);
+
 #endif /* CGIMAP_BACKEND_APIDB_COMMON_PGSQL_SELECTION_HPP */
